+++ CHANGELOG.md
@@ -12,10 +12,12 @@ Started since version 0.9.5
 
 * add the ability to hide action header (hideActionHeader)
 * add option to change the default href for a resource in a sidebar.
+* add avatarUrl to CurrentAdmin
 
 ### Fixed
 
 * fixed show/hide filter button
+* fix font from Theme not being passed to the frontend.
 
 ### Changed